feat: transaction endpoint

+++ src/main/java/dio/budgeting/application/input/PersistTransactionInput.java
@@ -2,6 +2,6 @@ package dio.budgeting.application.input;
 
 import dio.budgeting.domain.Category;
 
-public record PersistTransactionInput(String description, long amount, Category category) {
+public record PersistTransactionInput(String description, double amount, Category category) {
     
 }
