+++ src/main/java/dio/budgeting/infrastructure/http/request/TransactionRequest.java
@@ -3,7 +3,7 @@ package dio.budgeting.infrastructure.http.request;
 import dio.budgeting.application.input.PersistTransactionInput;
 import dio.budgeting.domain.Category;
 
-public record TransactionRequest(String description, Category category, long amount) {
+public record TransactionRequest(String description, Category category, double amount) {
 
     public PersistTransactionInput toInput() {
         return new PersistTransactionInput(description, amount, category);
